refactor: 상속 구조 변경
person을 상속받은 player를 person과 합침
