+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,6 +1,6 @@
 package blackjack.domain;
 
-public class Dealer extends Person{
+public class Dealer extends Player {
 
     private static final int CONDITION_HIT = 16;
     private static final String NAME = "딜러";
